res:dock:: change applet icon from dock.
1. switch mode.
2. switch theme icon.

tower:https://tower.im/projects/8f19f0bf0e754f0b82ef2c24bc230973/todos/2e5603ced09048dc8eb74f44c3c5d553/
tower:https://tower.im/projects/8f19f0bf0e754f0b82ef2c24bc230973/todos/5ef98cd230244b44b2236a8510a6f3a8/?proto=https

+++ lib/jsc_extension/dcore.c
@@ -52,7 +52,7 @@ enum PluginState {
 JS_EXPORT_API
 char* dcore_get_theme_icon(const char* name, double size)
 {
-    return icon_name_to_path_with_check_xpm(name, size);
+    return icon_name_to_path_with_check_xpm(name, (int)size);
 }
